requirements.txt
pandas
python-binance
requests